My First Commit on GitHub
This is code I developed to make my job automated and improve my workflow, as well as efficiency. It takes the client's cell phone number and formats it correctly and then opens it automatically in a new tab followed by the WhatsApp application opening, making it easier for you to start typing your message immediately. The code is made for pasting of cell phone numbers rather than entering it manually.